added possibility to change the API key and added new folder for ui files
