+++ src/main.py
@@ -6,6 +6,6 @@ if __name__ == "__main__":
     mode = setup_environment()
     print(f"Using {mode} mode")
 
-    from user_interface import AityApp
+    from ui.user_interface import AityApp
     app = AityApp()
     app.mainloop()
