+++ src/output/summary/Balkan trip (1).json
@@ -1,0 +1,8 @@
+{
+  "summary": "This document outlines a comprehensive Western Balkans travel itinerary for May and June 2026. It compares two transport options: a driving route starting from Finland and a bus-based route. Key stops include Belgrade, Sarajevo, Dubrovnik, Kotor, and Tirana, featuring specific dates, logistics, and accommodation recommendations for Shköder, Albania.",
+  "keywords": "Western Balkans, travel itinerary, road trip, public transport, 2026",
+  "topics": "Travel Planning, Balkan Tourism, Transport Logistics",
+  "token_count": 2049,
+  "mode": "genai",
+  "source_file": "C:\\Users\\Pihla\\Downloads\\Balkan trip (1).txt"
+}
