+++ src/output/summary/Luento 7 reflektio (2).json
@@ -1,0 +1,8 @@
+{
+  "summary": "Teksti käsittelee laadunhallintajärjestelmiä, ISO 9001 -standardia ja jatkuvaa kehittämistä. Laadunhallinta on organisaation kokonaisvaltaista laatutoimintaa, jota kehitetään muun muassa Lean- ja Six Sigma -malleilla. Tilastollinen prosessinohjaus auttaa tunnistamaan virheiden syitä ja saavuttamaan tilastollisen hallinnan ennen prosessien säätämistä tai jatkokehitystä.",
+  "keywords": "laadunhallintajärjestelmä, ISO 9001, jatkuva kehittäminen, Lean, Six Sigma",
+  "topics": "laadunhallinta, prosessien kehittäminen, tilastollinen prosessinohjaus",
+  "token_count": 3238,
+  "mode": "genai",
+  "source_file": "C:\\Users\\Pihla\\Downloads\\Luento 7 reflektio (2).pdf"
+}
