+++ src/ui/__init__.py
@@ -1,0 +1,1 @@
+# this empty file helps with folder structure
