+++ src/user_data.json
@@ -1,6 +1,6 @@
 {
     "api key": "",
-    "api_key": "AIzaSyAb4x4KhCUPUZ7cpstw5SKne6FpmoSjt6I",
-    "analysis_mode": "genai",
+    "api_key": "AIzaSyBq5OvKUHPoRLJ2lYSRD8Y2ln20msCPG8k",
+    "analysis_mode": "",
     "settings": {}
 }
